extract TagTranslator to interface
