+++ oshdb-filter/src/test/java/org/heigit/ohsome/oshdb/filter/FilterTest.java
@@ -1,6 +1,7 @@
 package org.heigit.ohsome.oshdb.filter;
 
 import java.io.IOException;
+import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -19,6 +20,7 @@ import org.heigit.ohsome.oshdb.osm.OSMRelation;
 import org.heigit.ohsome.oshdb.osm.OSMType;
 import org.heigit.ohsome.oshdb.osm.OSMWay;
 import org.heigit.ohsome.oshdb.util.exceptions.OSHDBKeytablesNotFoundException;
+import org.heigit.ohsome.oshdb.util.tagtranslator.DefaultTagTranslator;
 import org.heigit.ohsome.oshdb.util.tagtranslator.TagTranslator;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -27,22 +29,23 @@ import org.junit.jupiter.api.BeforeEach;
  * Tests the parsing of filters and the application to OSM entities.
  */
 abstract class FilterTest {
+  protected Connection conn;
   protected FilterParser parser;
   protected TagTranslator tagTranslator;
 
   @BeforeEach
   void setup() throws SQLException, ClassNotFoundException, OSHDBKeytablesNotFoundException {
-    Class.forName("org.h2.Driver");
-    this.tagTranslator = new TagTranslator(DriverManager.getConnection(
-        "jdbc:h2:../data/test-data;ACCESS_MODE_DATA=r",
-        "sa", ""
-    ));
+    this.conn = DriverManager.getConnection(
+            "jdbc:h2:../data/test-data;ACCESS_MODE_DATA=r",
+            "sa", ""
+    );
+    this.tagTranslator = new DefaultTagTranslator(conn);
     this.parser = new FilterParser(this.tagTranslator);
   }
 
   @AfterEach
   void teardown() throws SQLException {
-    this.tagTranslator.getConnection().close();
+    conn.close();
   }
 
   protected int[] createTestTags(String... keyValues) {
