+++ oshdb-util/src/main/java/org/heigit/ohsome/oshdb/util/taginterpreter/DefaultTagInterpreter.java
@@ -10,6 +10,7 @@ import java.util.TreeMap;
 import java.util.TreeSet;
 import org.heigit.ohsome.oshdb.osm.OSMEntity;
 import org.heigit.ohsome.oshdb.osm.OSMRelation;
+import org.heigit.ohsome.oshdb.util.tagtranslator.DefaultTagTranslator;
 import org.heigit.ohsome.oshdb.util.tagtranslator.OSMTag;
 import org.heigit.ohsome.oshdb.util.tagtranslator.TagTranslator;
 import org.json.simple.JSONArray;
@@ -35,7 +36,7 @@ public class DefaultTagInterpreter extends BaseTagInterpreter {
       = "json/uninterestingTags.json";
 
   /**
-   * Constructor using given {@link TagTranslator} and default values as areaTagsDefinitonFile and
+   * Constructor using given {@link DefaultTagTranslator} and default values as areaTagsDefinitonFile and
    * uninterestingTagsDefinitionFile.
    *
    * <p>
@@ -52,10 +53,10 @@ public class DefaultTagInterpreter extends BaseTagInterpreter {
   }
 
   /**
-   * Constructor using given {@link TagTranslator}, areaTagsDefinitonFile, and
+   * Constructor using given {@link DefaultTagTranslator}, areaTagsDefinitonFile, and
    * uninterestingTagsDefinitionFile.
    *
-   * @param tagTranslator {@link TagTranslator} used by {@link TagInterpreter}
+   * @param tagTranslator {@link DefaultTagTranslator} used by {@link TagInterpreter}
    * @param areaTagsDefinitionFile filename of a JSON file containing tags that are supposed to be
    *                               areas
    * @param uninterestingTagsDefinitionFile filename of a JSON file containing tags to be ignored
